+++ lib/turbopuffer.ex
@@ -51,9 +51,11 @@ defmodule Turbopuffer do
           | {:finch_name, atom()}
         ]
 
+  # Passed to Finch.request/3, see Turbopuffer.Client.request/5
   @type request_opts :: [
-          {:timeout, pos_integer()}
-          | {:receive_timeout, pos_integer()}
+          {:pool_timeout, timeout()}
+          | {:receive_timeout, timeout()}
+          | {:request_timeout, timeout()}
         ]
 
   # Filter types